+++ mednafen/wswan/sound.c
@@ -32,31 +32,31 @@ static Blip_Synth VoiceSynth;
 
 static Blip_Buffer sbuf[2];
 
-static uint16 period[4];
-static uint8 volume[4]; // left volume in upper 4 bits, right in lower 4 bits
-static uint8 voice_volume;
+static uint16_t period[4];
+static uint8_t volume[4]; // left volume in upper 4 bits, right in lower 4 bits
+static uint8_t voice_volume;
 
-static uint8 sweep_step, sweep_value;
-static uint8 noise_control;
-static uint8 control;
-static uint8 output_control;
+static uint8_t sweep_step, sweep_value;
+static uint8_t noise_control;
+static uint8_t control;
+static uint8_t output_control;
 
-static int32 sweep_8192_divider;
-static uint8 sweep_counter;
-static uint8 SampleRAMPos;
+static int32_t sweep_8192_divider;
+static uint8_t sweep_counter;
+static uint8_t SampleRAMPos;
 
-static int32 sample_cache[4][2];
+static int32_t sample_cache[4][2];
 
-static int32 last_v_val;
+static int32_t last_v_val;
 
-static uint8 HyperVoice;
-static int32 last_hv_val;
+static uint8_t HyperVoice;
+static int32_t last_hv_val;
 
-static int32 period_counter[4];
-static int32 last_val[4][2]; // Last outputted value, l&r
-static uint8 sample_pos[4];
-static uint16 nreg;
-static uint32 last_ts;
+static int32_t period_counter[4];
+static int32_t last_val[4][2]; // Last outputted value, l&r
+static uint8_t sample_pos[4];
+static uint16_t nreg;
+static uint32_t last_ts;
 
 
 #define MK_SAMPLE_CACHE	\
@@ -78,7 +78,7 @@ static uint32 last_ts;
 
 #define SYNCSAMPLE(wt)	\
    {	\
-    int32 left = sample_cache[ch][0], right = sample_cache[ch][1];	\
+    int32_t left = sample_cache[ch][0], right = sample_cache[ch][1];	\
     Blip_Synth_offset(&WaveSynth, wt, left - last_val[ch][0], &sbuf[0]);	\
     Blip_Synth_offset(&WaveSynth, wt, right - last_val[ch][1], &sbuf[1]);	\
     last_val[ch][0] = left;	\
@@ -87,7 +87,7 @@ static uint32 last_ts;
 
 #define SYNCSAMPLE_NOISE(wt)  \
    {    \
-    int32 left = sample_cache[ch][0], right = sample_cache[ch][1];      \
+    int32_t left = sample_cache[ch][0], right = sample_cache[ch][1];      \
     Blip_Synth_offset(&NoiseSynth, wt, left - last_val[ch][0], &sbuf[0]);      \
     Blip_Synth_offset(&NoiseSynth, wt, right - last_val[ch][1], &sbuf[1]);     \
     last_val[ch][0] = left;     \
@@ -96,7 +96,7 @@ static uint32 last_ts;
 
 void WSwan_SoundUpdate(void)
 {
-   int32 run_time;
+   int32_t run_time;
 
    //printf("%d\n", v30mz_timestamp);
    //printf("%02x %02x\n", control, noise_control);
@@ -110,7 +110,7 @@ void WSwan_SoundUpdate(void)
 
       if(ch == 1 && (control & 0x20)) // Direct D/A mode?
       {
-         int32 neoval = (volume[ch] - 0x80) * voice_volume;
+         int32_t neoval = (volume[ch] - 0x80) * voice_volume;
 
          Blip_Synth_offset(&VoiceSynth, v30mz_timestamp, neoval - last_v_val, &sbuf[0]);
          Blip_Synth_offset(&VoiceSynth, v30mz_timestamp, neoval - last_v_val, &sbuf[1]);
@@ -119,13 +119,13 @@ void WSwan_SoundUpdate(void)
       }
       else if(ch == 2 && (control & 0x40) && sweep_value) // Sweep
       {
-         uint32 tmp_pt = 2048 - period[ch];
-         uint32 meow_timestamp = v30mz_timestamp - run_time;
-         uint32 tmp_run_time = run_time;
+         uint32_t tmp_pt = 2048 - period[ch];
+         uint32_t meow_timestamp = v30mz_timestamp - run_time;
+         uint32_t tmp_run_time = run_time;
 
          while(tmp_run_time)
          {
-            int32 sub_run_time = tmp_run_time;
+            int32_t sub_run_time = tmp_run_time;
 
             if(sub_run_time > sweep_8192_divider)
                sub_run_time = sweep_8192_divider;
@@ -138,7 +138,7 @@ void WSwan_SoundUpdate(void)
                if(sweep_counter <= 0)
                {
                   sweep_counter = sweep_step + 1;
-                  period[ch] = (period[ch] + (int8)sweep_value) & 0x7FF;
+                  period[ch] = (period[ch] + (int8_t)sweep_value) & 0x7FF;
                }
             }
 
@@ -160,7 +160,7 @@ void WSwan_SoundUpdate(void)
       }
       else if(ch == 3 && (noise_control & 0x10)) //(control & 0x80)) // Noise
       {
-         uint32 tmp_pt = 2048 - period[ch];
+         uint32_t tmp_pt = 2048 - period[ch];
 
          period_counter[ch] -= run_time;
          while(period_counter[ch] <= 0)
@@ -187,7 +187,7 @@ void WSwan_SoundUpdate(void)
       }
       else
       {
-         uint32 tmp_pt = 2048 - period[ch];
+         uint32_t tmp_pt = 2048 - period[ch];
 
          if(tmp_pt > 4)
          {
@@ -205,7 +205,7 @@ void WSwan_SoundUpdate(void)
    }
 
    {
-      int32 tmphv = HyperVoice;
+      int32_t tmphv = HyperVoice;
 
       if(tmphv - last_hv_val)
       {
@@ -217,7 +217,7 @@ void WSwan_SoundUpdate(void)
    last_ts = v30mz_timestamp;
 }
 
-void WSwan_SoundWrite(uint32 A, uint8 V)
+void WSwan_SoundWrite(uint32_t A, uint8_t V)
 {
    WSwan_SoundUpdate();
 
@@ -282,7 +282,7 @@ void WSwan_SoundWrite(uint32 A, uint8 V)
    WSwan_SoundUpdate();
 }
 
-uint8 WSwan_SoundRead(uint32 A)
+uint8_t WSwan_SoundRead(uint32_t A)
 {
    WSwan_SoundUpdate();
 
@@ -317,9 +317,9 @@ uint8 WSwan_SoundRead(uint32 A)
 }
 
 
-int32 WSwan_SoundFlush(int16 *SoundBuf, const int32 MaxSoundFrames)
+int32_t WSwan_SoundFlush(int16_t *SoundBuf, const int32_t MaxSoundFrames)
 {
-   int32 FrameCount = 0;
+   int32_t FrameCount = 0;
 
    WSwan_SoundUpdate();
 
@@ -338,7 +338,7 @@ int32 WSwan_SoundFlush(int16 *SoundBuf, const int32 MaxSoundFrames)
 }
 
 // Call before wsRAM is updated
-void WSwan_SoundCheckRAMWrite(uint32 A)
+void WSwan_SoundCheckRAMWrite(uint32_t A)
 {
    if((A >> 6) == SampleRAMPos)
       WSwan_SoundUpdate();
@@ -346,7 +346,7 @@ void WSwan_SoundCheckRAMWrite(uint32 A)
 
 static void RedoVolume(void)
 {
-	double eff_volume = 1.0 / 4;
+	float eff_volume = 1.0f / 4;
 
 	Blip_Synth_set_volume(&WaveSynth, eff_volume, 256);
 	Blip_Synth_set_volume(&NoiseSynth, eff_volume, 256);
@@ -376,7 +376,7 @@ void WSwan_SoundKill(void)
 
 }
 
-uint32_t WSwan_SetSoundRate(uint32 rate)
+uint32_t WSwan_SetSoundRate(uint32_t rate)
 {
 	unsigned i;
 	for(i = 0; i < 2; i++)
