+++ mednafen/wswan/sound.h
@@ -3,17 +3,17 @@
 
 #include <stdint.h>
 
-int32 WSwan_SoundFlush(int16 *SoundBuf, const int32 MaxSoundFrames);
+int32_t WSwan_SoundFlush(int16_t *SoundBuf, const int32_t MaxSoundFrames);
 
 void WSwan_SoundInit(void);
 void WSwan_SoundKill(void);
-void WSwan_SetSoundMultiplier(double multiplier);
-uint32_t WSwan_SetSoundRate(uint32 rate);
+void WSwan_SetSoundMultiplier(float multiplier);
+uint32_t WSwan_SetSoundRate(uint32_t rate);
 
-void WSwan_SoundWrite(uint32, uint8);
-uint8 WSwan_SoundRead(uint32);
+void WSwan_SoundWrite(uint32_t, uint8_t);
+uint8_t WSwan_SoundRead(uint32_t);
 void WSwan_SoundReset(void);
-void WSwan_SoundCheckRAMWrite(uint32 A);
+void WSwan_SoundCheckRAMWrite(uint32_t A);
 
 void WSwan_SoundSaveState(uint32_t load, FILE* fp);
